HS-960: Move SnmpPeerFactory to shared-lib
