+++ shared-lib/horizon-common-utils/src/main/java/org/opennms/horizon/shared/utils/IPAddressRangeSet.java
@@ -26,7 +26,7 @@
  *     http://www.opennms.com/
  *******************************************************************************/
 
-package org.opennms.horizon.minion.snmp;
+package org.opennms.horizon.shared.utils;
 
 import java.util.Collections;
 import java.util.Iterator;
