+++ shared-lib/snmp/pom.xml
@@ -14,9 +14,10 @@
   <packaging>pom</packaging>
 
   <modules>
-    <module>snmp-api</module>
-    <module>snmp-impl</module>
+      <module>snmp-api</module>
+      <module>snmp-impl</module>
       <module>proto</module>
+      <module>agent-config</module>
   </modules>
 
 </project>
